feat: create board manager on server ready
